external modules: connections refactoring
